backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
python-multipart
transformers
torch
sentencepiece
google-genai
langdetect